+++ modules/s2/redshift/sql/ST_BOUNDARY.sql
@@ -2,10 +2,9 @@
 -- Copyright (C) 2021 CARTO
 ----------------------------
 
-CREATE OR REPLACE FUNCTION @@RS_PREFIX@@s2.ST_BOUNDARY(
-    id INT8
-) 
-RETURNS VARCHAR(MAX) 
+CREATE OR REPLACE FUNCTION @@RS_PREFIX@@s2.ST_BOUNDARY
+(id INT8) 
+RETURNS VARCHAR(MAX)
 IMMUTABLE
 AS $$
     from @@RS_PREFIX@@s2Lib import get_cell_boundary